Read and Combine Parquet Files into a Single Data Frame

This function reads all Parquet files from a specified S3 bucket and combines them into a single data frame.

CABCMParquetRead(
  sub = "cabcm",
  bucket = "tnc-dangermond",
  prefix = "water_balance/v2/"
)

Arguments

sub

A character string specifying the subfolder within the base folder to look for Parquet files. Default is "cabcm".

bucket

A string representing the S3 bucket name.

prefix

A string representing the prefix (path) within the S3 bucket where the Parquet files are stored.

Value

A data frame containing the combined data from all Parquet files in the specified subfolder. Each row in the combined data frame has an additional column named `file` indicating the source file of the data.

Details

The `CABCMParquetRead` function performs the following steps:

Lists all Parquet files in the specified subfolder within the base folder `"/Users/wamclean/Desktop/Lynker/tnc_hf/water_balance/"`.

Reads each Parquet file into a data frame and stores them in a list.

Combines all data frames in the list into a single data frame with an additional column `file` indicating the file from which each row was read.

Examples

# Read and combine all Parquet files from the "cabcm" subfolder
new_cabcm_data <- CABCMParquetRead()
#> List of 4
#>  $ Code     : chr "AccessDenied"
#>  $ Message  : chr "Access Denied"
#>  $ RequestId: chr "KN1KEW252KGVW4XK"
#>  $ HostId   : chr "5CFU8pItNST2oBfR1J7f8T+SmPf/XOdHh7Ij239p+ifm3qbgijQa3Isb+iwTgsRIDALjSMHpFxk="
#>  - attr(*, "headers")=List of 7
#>   ..$ x-amz-bucket-region: chr "us-west-2"
#>   ..$ x-amz-request-id   : chr "KN1KEW252KGVW4XK"
#>   ..$ x-amz-id-2         : chr "5CFU8pItNST2oBfR1J7f8T+SmPf/XOdHh7Ij239p+ifm3qbgijQa3Isb+iwTgsRIDALjSMHpFxk="
#>   ..$ content-type       : chr "application/xml"
#>   ..$ transfer-encoding  : chr "chunked"
#>   ..$ date               : chr "Thu, 10 Oct 2024 21:36:14 GMT"
#>   ..$ server             : chr "AmazonS3"
#>   ..- attr(*, "class")= chr [1:2] "insensitive" "list"
#>  - attr(*, "class")= chr "aws_error"
#> NULL
#> Error in parse_aws_s3_response(r, Sig, verbose = verbose): Forbidden (HTTP 403).

# Read and combine all Parquet files from a different subfolder
other_data <- CABCMParquetRead(sub = "another_subfolder")
#> List of 4
#>  $ Code     : chr "AccessDenied"
#>  $ Message  : chr "Access Denied"
#>  $ RequestId: chr "KN1KBSP38WN8Q52J"
#>  $ HostId   : chr "ro5ebOSxnLqZpwuE/NFhkqOR1HzAYl40yH8uicTjzXEzPqvPqY0mAq4anrXF5AB4zZww9G1an+w="
#>  - attr(*, "headers")=List of 7
#>   ..$ x-amz-bucket-region: chr "us-west-2"
#>   ..$ x-amz-request-id   : chr "KN1KBSP38WN8Q52J"
#>   ..$ x-amz-id-2         : chr "ro5ebOSxnLqZpwuE/NFhkqOR1HzAYl40yH8uicTjzXEzPqvPqY0mAq4anrXF5AB4zZww9G1an+w="
#>   ..$ content-type       : chr "application/xml"
#>   ..$ transfer-encoding  : chr "chunked"
#>   ..$ date               : chr "Thu, 10 Oct 2024 21:36:14 GMT"
#>   ..$ server             : chr "AmazonS3"
#>   ..- attr(*, "class")= chr [1:2] "insensitive" "list"
#>  - attr(*, "class")= chr "aws_error"
#> NULL
#> Error in parse_aws_s3_response(r, Sig, verbose = verbose): Forbidden (HTTP 403).