Plot Seasons

This function generates standardized plots for the Dangermond Preserve based on seasonal data. The plots visualize the percent error across different spatial features for a given season. This function is designed to be used within the `GridSeasons()` function to arrange the plots in a grid layout.

PlotSeason(SeasonData, season_name)

Arguments

SeasonData

A spatial dataframe containing seasonal data with columns such as `percent_error` and `type`. This data should be output from the `process_model_data()` function.

season_name

A character string representing the name of the season (e.g., "Winter", "Spring", "Summer", "Fall"). This will be used as the title of the plot.

Value

A `ggplot` object displaying the percent error across the spatial features of the Dangermond Preserve for the specified season.

Details

The `PlotSeason` function performs the following operations:

**Plot Creation**: Generates a ggplot object with a map tile background, using OpenStreetMap data.

**Coordinate System**: Uses the spatial coordinate reference system `st_crs(26910)`.

**Visualization**: Plots the spatial data with `percent_error` values visualized using a gradient color scale from white to dark red.

**Title**: Sets the plot title to the provided `season_name`.