Process Data for a List of Divide IDs

This function processes data for a list of divide IDs by selecting the corresponding basins and reading the associated parquet files from S3. The data is combined into a single data frame.

Usage

process_divide_ids(divide_ids, bucket_name, prefix, region)

Arguments

divide_ids

A character vector containing the divide IDs to process.

bucket_name

A character string representing the name of the S3 bucket.

prefix

A character string representing the prefix path in the S3 bucket.

region

A character string representing the AWS region. Default is `"us-west-2"`.

Value

A data frame containing the combined data for all the divide IDs.