Read Parquet Files from S3

This function lists and reads parquet files from an S3 bucket for a given set of site names. The data is processed to group by date and summarize numeric columns.

Usage

read_parquet_files(site_names, bucket_name, prefix, region = "us-west-2")

Arguments

site_names

A character vector containing site names for which to read parquet files.

bucket_name

A character string representing the name of the S3 bucket.

prefix

A character string representing the prefix path in the S3 bucket.

region

A character string representing the AWS region. Default is `"us-west-2"`.

Value

A data frame containing the combined data from the parquet files, grouped by date.